x-pack/filebeat/input/internal: split httpmon out of httplog (#36385)
